+++ src/com/jpexs/decompiler/flash/gui/FontFace.java
@@ -44,6 +44,14 @@ public class FontFace implements Comparable<FontFace> {
         return face;
     }
 
+    public boolean isBold() {
+        return toString().toLowerCase().contains("bold");
+    }
+    
+    public boolean isItalic() {
+        return toString().toLowerCase().contains("italic");
+    }
+    
     @Override
     public int hashCode() {
         int hash = 7;